pyproject.toml
[project]
name = "nonebot-plugin-werewolf"
version = "1.1.1"
description = "适用于 Nonebot2 的狼人杀插件"
authors = [{ name = "wyf7685", email = "wyf7685@163.com" }]
dependencies = [
    "nonebot2>=2.3.3",
    "nonebot-plugin-alconna>=0.52.1",
    "nonebot-plugin-userinfo>=0.2.6",
    "nonebot-plugin-waiter>=0.7.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[tool.uv]
dev-dependencies = [
    "nonebot2[fastapi]>=2.3.3",
    "nonebot-adapter-onebot>=2.4.4",
    "nonebot-adapter-satori>=0.12.5",
    "fastapi>=0.114.0",
    "httpx>=0.27.2",
    "websockets>=13.0.1",
]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "RUF",   # Ruff-specific rules
    "ERA",   # eradicate
    "FLY",   # flynt
    "FURB",  # refurb
    "I",     # isort
    "N",     # PEP8-naming
    "PD",    # pandas-vet
    "PERF",  # Perflint
    "PGH",   # pygrep-hooks
    "TRY",   # tryceratops
    "ASYNC", # flake8-async
    "ANN",   # flake8-annotations
    "S",     # flake8-bandit
    "FBT",   # flake8-boolean-trap
    "A",     # flake8-builtins
    "B",     # flake8-bugbear
    "COM",   # flake8-commas
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T10",   # flake8-debugger
    "EM",    # flake8-errmsg
    "FA",    # flake8-future-annotations
    "ISC",   # flake8-implicit-str-concat
    "ICN",   # flake8-import-conventions
    "PIE",   # flake8-pie
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "SLF",   # flake8-self
    "SIM",   # flake8-simplify
    "SLOT",  # flake8-slots
    "TID",   # flake8-tidy-imports
    "ARG",   # flake8-unused-arguments
    "PTH",   # flake8-use-pathlib
    "TCH",   # flake8-type-checking
    "YTT",   # flake8-2020
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
    "ANN101", # missing-type-self
    "ANN102", # missing-type-cls
    "TRY003", # raise-vanilla-args
    "EM102",  # f-string-in-exception
    "COM812", # missing-trailing-comma
    "ERA001", # commented-out-code
    "EM101",  # raw-string-in-exception
]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "standard"
reportShadowedImports = false
disableBytesTypePromotions = true
